Terminado

+++ src/main/java/ui/Details.java
@@ -4,9 +4,11 @@ import context.ContextService;
 import data.CsvDataService;
 import data.DataService;
 import data.Pelicula;
+import user.Usuario;
 
 import javax.swing.*;
 import java.awt.event.*;
+import java.util.Optional;
 
 public class Details extends JDialog {
     private JPanel contentPane;
@@ -23,6 +25,8 @@ public class Details extends JDialog {
 
 
     public Details(JFrame parent) {
+
+        this.dataService = new CsvDataService("peliculas.csv");
         setContentPane(contentPane);
         setModal(true);
         getRootPane().setDefaultButton(buttonOK);
@@ -58,8 +62,20 @@ public class Details extends JDialog {
 
         eliminarButton.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                dataService.delete();
-                dispose();
+                if (ContextService.getInstance().getItem("usuarioActivo").isPresent()) {
+                    Optional<Object> peliculaOpt = ContextService.getInstance().getItem("peliculaSeleccionada");
+                    Pelicula peliculaEliminar = (Pelicula) peliculaOpt.get();
+                    int usuarioPelicula = peliculaEliminar.getId_UsuarioPelicula();
+                    Optional<Object> usuarioOpt = ContextService.getInstance().getItem("usuarioActivo");
+                    Usuario usuario = (Usuario) usuarioOpt.get();
+                    int idUsuario = usuario.getId_Usuario();
+                    if (idUsuario == usuarioPelicula) {
+                        dataService.delete();
+                        dispose();
+                    } else
+                        JOptionPane.showMessageDialog(null, "La pelicula solo la puede eliminar el usuario que lo ha creado, la id del usuario que la creó es " + usuarioPelicula, "", JOptionPane.WARNING_MESSAGE);
+                } else
+                    JOptionPane.showMessageDialog(null, "Necesitas iniciar sesion para eliminar una pelicula", null, JOptionPane.WARNING_MESSAGE);
             }
         });
     }
